chore: Refactoring graphql functions
