+++ packages/bff/src/entities.ts
@@ -10,8 +10,6 @@ import {
   UpdateDateColumn,
 } from 'typeorm';
 
-import { ProfileRepository } from './db.ts';
-
 @Entity({ name: 'profile' })
 export class ProfileTable {
   @PrimaryColumn()
@@ -67,26 +65,3 @@ export class SavedSearch {
   @DeleteDateColumn()
   deletedAt: Date;
 }
-
-export const getOrCreateProfile = async (sub: string, locale: string): Promise<ProfileTable> => {
-  const profile = await ProfileRepository!.findOne({
-    where: { sub },
-  });
-
-  if (!profile) {
-    const newProfile = new ProfileTable();
-    newProfile.sub = sub;
-    newProfile.language = locale || 'nb';
-
-    const savedProfile = await ProfileRepository!.save(newProfile);
-    if (!savedProfile) {
-      throw new Error('Fatal: Not able to create new profile');
-    }
-    return savedProfile;
-  }
-  if (profile?.language !== locale) {
-    profile.language = locale;
-    await ProfileRepository!.save(profile);
-  }
-  return profile;
-};
